Clarify remote assets controller is a Crafter Delivery feature
The ability to reference remote assets is a feature provided by Crafter Delivery (not Studio). Remote asset data source is a Crafter Studio feature that helps create the URL patterns mentioned here. It is the remote assets controller of Crafter Delivery that parses that URL pattern and provides the access to browsers.